refactor: 모키스트 스타일의 테스트 정리 및 관련 추상 클래스 제거.
