--- src/test/java/kitchenpos/DatabaseCleanup.java
@@ -1,40 +1,0 @@
-package kitchenpos;
-
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.stereotype.Component;
-
-import javax.persistence.Entity;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Table;
-import javax.transaction.Transactional;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Component
-public class DatabaseCleanup implements InitializingBean {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    private List<String> tableNames;
-
-    @Override
-    public void afterPropertiesSet() {
-        tableNames = entityManager.getMetamodel().getEntities().stream()
-                .filter(entity -> entity.getJavaType().getAnnotation(Entity.class) != null)
-                .map(entity -> entity.getJavaType().getAnnotation(Table.class).name())
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void execute() {
-        entityManager.flush();
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
-        for (String tableName : tableNames) {
-            entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-        }
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
-    }
-
-}
