--- src/test/java/kitchenpos/UnitTest.java
@@ -1,8 +1,0 @@
-package kitchenpos;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public class UnitTest {
-}
