+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,47 +1,49 @@
 package kitchenpos.application;
 
-import kitchenpos.UnitTest;
+import kitchenpos.IntegrationTest;
 import kitchenpos.domain.*;
 import kitchenpos.fixture.MenuFixture;
 import kitchenpos.fixture.MenuGroupFixture;
+import kitchenpos.fixture.MenuProductFixture;
 import kitchenpos.fixture.ProductFixture;
 import kitchenpos.infra.PurgomalumClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.test.mock.mockito.MockBean;
 
 import java.math.BigDecimal;
 import java.util.*;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
+import static org.mockito.BDDMockito.given;
 
-@ExtendWith(MockitoExtension.class)
-class MenuServiceTest extends UnitTest {
 
-    @Mock
-    private MenuRepository menuRepository;
+class MenuServiceTest extends IntegrationTest {
 
-    @Mock
-    private MenuGroupRepository menuGroupRepository;
+    private final MenuService menuService;
 
-    @Mock
-    private ProductRepository productRepository;
+    private final MenuRepository menuRepository;
 
-    @Mock
+    private final MenuGroupRepository menuGroupRepository;
+
+    private final ProductRepository productRepository;
+
+    @MockBean
     private PurgomalumClient purgomalumClient;
 
-    @InjectMocks
-    private MenuService menuService;
+    MenuServiceTest(MenuService menuService, MenuRepository menuRepository, MenuGroupRepository menuGroupRepository, ProductRepository productRepository) {
+        this.menuService = menuService;
+        this.menuRepository = menuRepository;
+        this.menuGroupRepository = menuGroupRepository;
+        this.productRepository = productRepository;
+    }
+
 
     static class create_source {
         static Object[][] create_fail_because_illegal_price() {
@@ -59,22 +61,19 @@ class MenuServiceTest extends UnitTest {
         @DisplayName("[정상] 메뉴가 정상적으로 등록됩니다.")
         @Test
         void create_success() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            given(purgomalumClient.containsProfanity(any())).willReturn(false);
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(19_000L),
+                    menuProducts, menuGroup, true
             );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(19_000L), products);
 
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-            when(productRepository.findAllByIdIn(any())).thenReturn(products);
-            products.forEach(
-                product -> when(productRepository.findById(product.getId()))
-                    .thenReturn(Optional.of(product))
-            );
-            when(purgomalumClient.containsProfanity(any())).thenReturn(false);
-            when(menuRepository.save(any())).thenReturn(menu);
-
-            menuService.create(menu);
+            menuService.create(creatingMenu);
         }
 
         @DisplayName("[예외] 등록 요청한 메뉴의 가격이 null 이거나 0원 미만 입니다.")
@@ -88,11 +87,13 @@ class MenuServiceTest extends UnitTest {
         @DisplayName("[예외] 등록 요청한 메뉴의 메뉴 구성 상품이 존재하지 않습니다.")
         @Test
         void create_fail_because_menu_products_do_not_exist() {
-            Menu menu = MenuFixture.create("후라이드 치킨 세트", BigDecimal.valueOf(19_000L), new ArrayList<>());
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(19_000L),
+                    new ArrayList<>(), menuGroup, true
+            );
 
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-
-            assertThatThrownBy(() -> menuService.create(menu))
+            assertThatThrownBy(() -> menuService.create(creatingMenu))
                 .isInstanceOf(IllegalArgumentException.class);
         }
 
@@ -100,79 +101,72 @@ class MenuServiceTest extends UnitTest {
         @Test
         void create_fail_because_some_product_do_not_exist() {
             List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
             );
-            Product firstProduct = products.get(0);
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(19_000L), products);
+            productRepository.save(products.get(0));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(19_000L),
+                    menuProducts, menuGroup, true
+            );
 
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-            when(productRepository.findAllByIdIn(any())).thenReturn(List.of(firstProduct));
-
-            assertThatThrownBy(() -> menuService.create(menu))
+            assertThatThrownBy(() -> menuService.create(creatingMenu))
                 .isInstanceOf(IllegalArgumentException.class);
         }
 
         @DisplayName("[예외] 메뉴를 구성하는 일부 메뉴 구성 상품의 개수가 0개 미만 입니다.")
         @Test
         void create_fail_because_some_quantity_of_menu_product_is_minus() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(19_000L), products);
-            menu.getMenuProducts().get(1).setQuantity(0L);
-
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-            when(productRepository.findAllByIdIn(any())).thenReturn(products);
-            products.forEach(
-                product -> when(productRepository.findById(product.getId()))
-                    .thenReturn(Optional.of(product))
+            Product product = productRepository.save(ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)));
+            MenuProduct menuProduct = MenuProductFixture.create(product, -1);
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(19_000L),
+                    List.of(menuProduct), menuGroup, true
             );
 
-            assertThatThrownBy(() -> menuService.create(menu))
+            assertThatThrownBy(() -> menuService.create(creatingMenu))
                 .isInstanceOf(IllegalArgumentException.class);
         }
 
         @DisplayName("[예외] 메뉴의 가격의 `sum(메뉴를 구성하는 상품의 가격 * 개수)` 보다 큽니다.")
         @Test
         void create_fail_because_menu_price_is_over_than_sum_of_menu_product_price() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(20_001L), products);
-
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-            when(productRepository.findAllByIdIn(any())).thenReturn(products);
-            products.forEach(
-                product -> when(productRepository.findById(product.getId()))
-                    .thenReturn(Optional.of(product))
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(20_001L),
+                    menuProducts, menuGroup, true
             );
 
-            assertThatThrownBy(() -> menuService.create(menu))
+            assertThatThrownBy(() -> menuService.create(creatingMenu))
                 .isInstanceOf(IllegalArgumentException.class);
         }
 
         @DisplayName("[예외] 메뉴의 이름에 비속어가 포함되어서는 안됩니다.")
         @Test
         void create_fail_because_menu_name_is_purgomalum() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            Menu creatingMenu = MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(19_001L),
+                    menuProducts, menuGroup, true
             );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(19_000L), products);
 
-            when(menuGroupRepository.findById(any())).thenReturn(Optional.of(menu.getMenuGroup()));
-            when(productRepository.findAllByIdIn(any())).thenReturn(products);
-            products.forEach(
-                product -> when(productRepository.findById(product.getId()))
-                    .thenReturn(Optional.of(product))
-            );
-            when(purgomalumClient.containsProfanity(any())).thenReturn(true);
+            given(purgomalumClient.containsProfanity(any())).willReturn(true);
 
-            assertThatThrownBy(() -> menuService.create(menu))
-                    .isInstanceOf(IllegalArgumentException.class);
+            assertThatThrownBy(() -> menuService.create(creatingMenu))
+              .isInstanceOf(IllegalArgumentException.class);
         }
     }
 
@@ -205,17 +199,19 @@ class MenuServiceTest extends UnitTest {
         @MethodSource("kitchenpos.application.MenuServiceTest$changePrice_source#create_success")
         @ParameterizedTest
         void create_success(BigDecimal changedPrice) {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(1_000L), products);
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            Menu createdMenu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(1_000L),
+                    menuProducts, menuGroup, true
+            ));
 
-            Menu changingPriceMenu = MenuFixture.create(menu.getId(), changedPrice, menu.getMenuGroup());
-
-            when(menuRepository.findById(any())).thenReturn(Optional.of(menu));
-
-            menuService.changePrice(menu.getId(), changingPriceMenu);
+            Menu changingPriceMenu = MenuFixture.create(createdMenu.getId(), changedPrice, menuGroup);
+            menuService.changePrice(changingPriceMenu.getId(), changingPriceMenu);
         }
 
         @DisplayName("[예외] 변경되는 메뉴의 가격은 null 이거나 음수 일 수 없습니다.")
@@ -229,19 +225,19 @@ class MenuServiceTest extends UnitTest {
         @DisplayName("[예외] 변경되는 메뉴의 가격은 개별 구성 상품 가격 * 개수 보다 클 수 없습니다.")
         @Test
         void create_fail_because_menu_price_is_over_than_each_menu_product_price_multiply_quantity() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(1_000L), products);
-            MenuGroup menuGroup = MenuGroupFixture.create();
-            Menu changingPriceMenu = MenuFixture.create(
-                menu.getId(), "후라이드 치킨 세트", BigDecimal.valueOf(2_001L), menu.getMenuProducts(), menuGroup, true
-            );
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            Menu createdMenu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(1_000L),
+                    menuProducts, menuGroup, true
+            ));
 
-            when(menuRepository.findById(any())).thenReturn(Optional.of(menu));
-
-            assertThatThrownBy(() -> menuService.changePrice(menu.getId(), changingPriceMenu))
+            Menu changingPriceMenu = MenuFixture.create(createdMenu.getId(), BigDecimal.valueOf(2_001L), menuGroup);
+            assertThatThrownBy(() -> menuService.changePrice(changingPriceMenu.getId(), changingPriceMenu))
                 .isInstanceOf(IllegalArgumentException.class);
         }
     }
@@ -258,28 +254,36 @@ class MenuServiceTest extends UnitTest {
         @DisplayName("[정상] 메뉴가 정상적으로 전시됩니다.")
         @Test
         void display_success() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(2_000L), products);
-
-            when(menuRepository.findById(any())).thenReturn(Optional.of(menu));
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            given(purgomalumClient.containsProfanity(any())).willReturn(false);
+            Menu menu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(2000L),
+                    menuProducts, menuGroup, false
+            ));
 
             Menu actualResult = menuService.display(menu.getId());
             assertTrue(actualResult.isDisplayed());
         }
 
-        @DisplayName("[정상] 메뉴가 전시시 메뉴의 가격이 개별 구성 상품 가격 * 개수 보다 크면 전시가 불가합니다.")
+        @DisplayName("[예외] 메뉴가 전시시 메뉴의 가격이 개별 구성 상품 가격 * 개수 보다 크면 전시가 불가합니다.")
         @Test
         void display_fail_because_menu_price_is_over_than_each_menu_product_price_multiply_quantity() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(2_001L), products);
-
-            when(menuRepository.findById(any())).thenReturn(Optional.of(menu));
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            given(purgomalumClient.containsProfanity(any())).willReturn(false);
+            Menu menu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(2_001L),
+                    menuProducts, menuGroup, false
+            ));
 
             assertThatThrownBy(() -> menuService.display(menu.getId()))
                 .isInstanceOf(IllegalStateException.class);
@@ -292,17 +296,46 @@ class MenuServiceTest extends UnitTest {
         @DisplayName("[정상] 메뉴가 정상적으로 숨겨집니다.")
         @Test
         void hide_success() {
-            List<Product> products = Arrays.asList(
-                ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
-                ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
-            );
-            Menu menu = MenuFixture.createWithProducts("후라이드 치킨 세트", BigDecimal.valueOf(2_000L), products);
-
-            when(menuRepository.findById(any())).thenReturn(Optional.of(menu));
+            List<Product> products = productRepository.saveAll(Arrays.asList(
+                    ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+                    ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+            ));
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+            given(purgomalumClient.containsProfanity(any())).willReturn(false);
+            Menu menu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(2_001L),
+                    menuProducts, menuGroup, true
+            ));
 
             Menu actualResult = menuService.hide(menu.getId());
             assertFalse(actualResult.isDisplayed());
         }
+    }
+
+
+    @Test
+    @DisplayName("메뉴를 조회합니다.")
+    void findAll() {
+        List<Product> products = productRepository.saveAll(Arrays.asList(
+          ProductFixture.create(UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(18_000L)),
+          ProductFixture.create(UUID.randomUUID(), "코카콜라 1.5L", BigDecimal.valueOf(2_000L))
+        ));
+        MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+        List<MenuProduct> menuProducts = MenuProductFixture.create(products , 1);
+        given(purgomalumClient.containsProfanity(any())).willReturn(false);
+        Menu firstMenu = menuRepository.save(MenuFixture.create(
+          UUID.randomUUID(), "후라이드 치킨 세트", BigDecimal.valueOf(2_001L),
+          menuProducts, menuGroup, true
+        ));
+        Menu secondMenu = menuRepository.save(MenuFixture.create(
+          UUID.randomUUID(), "후라이드 치킨 세트2", BigDecimal.valueOf(2_001L),
+          menuProducts, menuGroup, true
+        ));
+
+        List<Menu> actualResult = menuService.findAll();
+
+        assertThat(actualResult).containsExactly(firstMenu, secondMenu);
     }
 
 }
