+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,11 +1,13 @@
 package kitchenpos.application;
 
-import kitchenpos.UnitTest;
+import kitchenpos.IntegrationTest;
 import kitchenpos.domain.*;
 import kitchenpos.fixture.MenuFixture;
+import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.MenuProductFixture;
 import kitchenpos.fixture.ProductFixture;
 import kitchenpos.infra.PurgomalumClient;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -13,35 +15,47 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
+import org.springframework.boot.test.mock.mockito.MockBean;
 
 import java.math.BigDecimal;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 import static java.math.BigDecimal.ZERO;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.when;
 
-class ProductServiceTest extends UnitTest {
+class ProductServiceTest extends IntegrationTest {
 
-    @Mock
-    private ProductRepository productRepository;
+    private final ProductService productService;
+    private final ProductRepository productRepository;
+    private final MenuRepository menuRepository;
+    private final MenuGroupRepository menuGroupRepository;
 
-    @Mock
-    private MenuRepository menuRepository;
+    @MockBean
+    private final PurgomalumClient purgomalumClient;
 
-    @Mock
-    private PurgomalumClient purgomalumClient;
 
-    @InjectMocks
-    private ProductService productService;
+    ProductServiceTest(ProductService productService,
+                       ProductRepository productRepository,
+                       MenuRepository menuRepository,
+                       MenuGroupRepository menuGroupRepository,
+                       PurgomalumClient purgomalumClient) {
+        this.productService = productService;
+        this.productRepository = productRepository;
+        this.menuRepository = menuRepository;
+        this.menuGroupRepository = menuGroupRepository;
+        this.purgomalumClient = purgomalumClient;
+    }
+    private Product 상품_만원;
+    @BeforeEach
+    void setUp() {
+        상품_만원 = ProductFixture.create(BigDecimal.valueOf(10_000L));
 
+    }
 
     public static class create_source {
         public static Object[][] source_of_create_success() {
@@ -54,24 +68,26 @@ class ProductServiceTest extends UnitTest {
         }
     }
 
-
     @DisplayName("상품을 등록합니다.")
     @Nested
     class create {
 
         @DisplayName("[정상] 상품이 정상적으로 등록됩니다.")
         @ParameterizedTest
-        @MethodSource("kitchenpos.application.ProductServiceTest$create_source#source_of_create_success")
+        @MethodSource("kitchenpos.application.ProductServiceIntegrationTest$create_source#source_of_create_success")
         void create_success(String name, BigDecimal price) {
             Product product = ProductFixture.create(name, price);
 
-            productService.create(product);
+            Product actualResult = productService.create(product);
+
+            assertEquals(product.getName(), actualResult.getName());
+            assertEquals(product.getPrice(), actualResult.getPrice());
         }
 
         @DisplayName("[예외] 상품의 이름은 비속어일 수 없다.")
         @Test
         void create_fail_because_profanity_name() {
-            Product product = ProductFixture.create(BigDecimal.valueOf(10000L));
+            Product product = 상품_만원;
             given(purgomalumClient.containsProfanity(product.getName())).willReturn(true);
 
             assertThatThrownBy(() -> productService.create(product))
@@ -82,7 +98,7 @@ class ProductServiceTest extends UnitTest {
         @NullSource
         @ParameterizedTest
         void create_fail_because_null_name(String name) {
-            Product product = ProductFixture.create(name, BigDecimal.valueOf(10000L));
+            Product product = ProductFixture.create(name, BigDecimal.valueOf(10_000L));
 
             assertThatThrownBy(() -> productService.create(product))
                     .isInstanceOf(IllegalArgumentException.class);
@@ -104,15 +120,11 @@ class ProductServiceTest extends UnitTest {
     public static class changePrice_source {
 
         public static Object[][] changePrice_success() {
-            Product product = ProductFixture.create(BigDecimal.valueOf(10000));
-            MenuProduct menuProduct = MenuProductFixture.create(product, 3);
-            Menu menu = MenuFixture.create("후라이드 치킨", BigDecimal.valueOf(30000), menuProduct);
-
             return new Object[][]{
-                    {"상품의 가격이 20,000원으로 오른 경우", product, menu, BigDecimal.valueOf(20000), true},
-                    {"상품의 가격이 10,000원으로 동일한 경우", product, menu, BigDecimal.valueOf(10000), true},
-                    {"상품의 가격이 8,000원으로 내려간 경우", product, menu, BigDecimal.valueOf(8000), false},
-                    {"상품의 가격이 0원으로 내려간 경우", product, menu, BigDecimal.valueOf(0), false},
+                    {"상품의 가격이 20,000원으로 오른 경우", BigDecimal.valueOf(20_000), true},
+                    {"상품의 가격이 10,000원으로 동일한 경우", BigDecimal.valueOf(10_000), true},
+                    {"상품의 가격이 8,000원으로 내려간 경우", BigDecimal.valueOf(8_000), false},
+                    {"상품의 가격이 0원으로 내려간 경우", BigDecimal.valueOf(0), false},
             };
         }
         public static Object[][] changePrice_fail_because_illegal_price() {
@@ -129,28 +141,42 @@ class ProductServiceTest extends UnitTest {
     class changePrice {
 
         @DisplayName("[정상] 상품의 가격을 변경합니다.")
-        @MethodSource("kitchenpos.application.ProductServiceTest$changePrice_source#changePrice_success")
+        @MethodSource("kitchenpos.application.ProductServiceIntegrationTest$changePrice_source#changePrice_success")
         @ParameterizedTest(name = "{0}")
-        void changePrice_success(String testName, Product product, Menu menu, BigDecimal changingPrice, boolean isDisplayed) {
-            when(productRepository.findById(any())).thenReturn(Optional.of(product));
-            when(menuRepository.findAllByProductId(any())).thenReturn(List.of(menu));
-
+        void changePrice_success(String testName, BigDecimal changingPrice, boolean isDisplayed) {
+            Product product = productRepository.save(상품_만원);
+            MenuProduct menuProduct = MenuProductFixture.create(product, 3);
+            MenuGroup menuGroup = menuGroupRepository.save(MenuGroupFixture.create());
+            Menu menu = menuRepository.save(MenuFixture.create(
+                    UUID.randomUUID(), "후라이드 치킨", BigDecimal.valueOf(30000),
+              List.of(menuProduct), menuGroup, true
+            ));
             Product changingProduct = ProductFixture.create(changingPrice);
 
-            productService.changePrice(UUID.randomUUID(), changingProduct);
+            Product actualResult = productService.changePrice(product.getId(), changingProduct);
 
-            assertEquals(changingProduct.getPrice(), product.getPrice());
+            assertEquals(changingProduct.getPrice(), actualResult.getPrice());
             assertEquals(menu.isDisplayed(), isDisplayed);
         }
 
         @DisplayName("[예외] 변경되는 상품의 가격은 null이거나 0미만 일 수 없습니다.")
-        @MethodSource("kitchenpos.application.ProductServiceTest$changePrice_source#changePrice_fail_because_illegal_price")
+        @MethodSource("kitchenpos.application.ProductServiceIntegrationTest$changePrice_source#changePrice_fail_because_illegal_price")
         @ParameterizedTest(name = "{0}")
         void changePrice_fail_because_illegal_price(String testName, Product product) {
             assertThatThrownBy(() -> productService.changePrice(UUID.randomUUID(), product))
                     .isInstanceOf(IllegalArgumentException.class);
         }
 
+    }
+
+    @Test
+    void findAll() {
+        Product chicken = productRepository.save(ProductFixture.create( "후라이드 치킨", BigDecimal.valueOf(18_000)));
+        Product coke = productRepository.save(ProductFixture.create("코카콜라", BigDecimal.valueOf(2_000)));
+
+        List<Product> actualResult = productService.findAll();
+
+        assertThat(actualResult).containsExactly(chicken, coke);
     }
 
 }
